Fix bad conversion of palette mode PNGs to grayscale
It occured when the palette was a candidate for conversion to grayscale,
and the source image uses the same grayscale level for both transparent
and opaque parts of the image. PngOptimizer would merge the two colors
into a single grayscale level.

Now, PngOptimizer will look for an unused gray level to be used as the
transparent gray level. For example if black is not used (level 0), the
image will contain black pixels for transparent parts.